+++ nrovalidate/validator/noderesourcetopologies.go
