+++ tools/nrtcacheck/nrtcacheck.go
